+++ packages/api/src/app/index.ts
@@ -1,4 +1,4 @@
-export * from './appConfigService';
+export * from './service';
 export * from './config';
 export * from './permissions';
 export * from './cdn';
